Add post_process_script_function method
Previously, there was no definition for the post_process_script_function method. This commit adds the method to the codebase, which allows for post-processing scripts to be run on preservation packages. The method takes two parameters: the path to the preservation package and the result code of the pre-processing script.