Add definition and parameters for 'post_process_script_function' method
Previously, there was no definition for the 'post_process_script_function' method in 'Article.py' file. This commit adds the method to the codebase, which allows for post-processing scripts to be run on preservation article packages. The method takes two parameters: the path to the preservation article package and the result code of the pre-processing script. Currently, the parameter 'result code of the pre-processing script' is being passed to the method but is not being utilized. It is planned to use this parameter in the future.