Calculate curation folder size based on matched curation folders
Currently, the required space is calculated by summing the curation folder size and the total size of all articles. The curation folder size is determined by considering all curation folders. However, when article IDs are provided explicitly, it is unnecessary to consider all curation folders. Instead, only the curation folder(s) that match the provided article ID(s) should be considered for calculating the 'curation folder size'. This refinement enables a more accurate calculation of the required space. The code responsible for this functionality has been developed and committed as part of this commit